requirements.txt
# LiveKit core and plugins
livekit-agents[deepgram,silero]==1.2.2
livekit-plugins-azure==1.2.2
livekit-plugins-google==1.2.2

# Env loading
python-dotenv>=1.0.1,<2.0.0

# Google LLM and Speech
google-cloud-speech>=2.26.0,<3.0.0
google-generativeai>=0.5.4,<1.0.0

# HTTP and utility libraries
aiohttp>=3.9.5,<4.0.0
requests>=2.31.0,<3.0.0
langdetect>=1.0.9,<2.0.0
soundfile>=0.12.1,<1.0.0

# ✅ Add these for smoother resolution
numpy>=1.24.0,<2.0.0